+++ models/product.go
@@ -5,11 +5,12 @@ type Product struct {
 	Name        string      `json:"name" gorm:"size:32;not null"`
 	Description string      `json:"description" gorm:"size:256;not null"`
 	Images      StringArray `json:"images"  gorm:"not null"`
-	Price       float64     `json:"price" gorm:"not null"`
-	Type        ProductType `json:"type" gorm:"not null"`
-	Closed      bool        `json:"closed" gorm:"not null"`
-	UserID      int         `json:"user_id" gorm:"index;not null"`
-	CategoryID  int         `json:"category_id" gorm:"not null"`
+	// Price in cent, $2.70 = 270
+	Price      int         `json:"price" gorm:"not null"`
+	Type       ProductType `json:"type" gorm:"not null"`
+	Closed     bool        `json:"closed" gorm:"not null"`
+	UserID     int         `json:"user_id" gorm:"index;not null"`
+	CategoryID int         `json:"category_id" gorm:"not null"`
 	// 由于目前租户较少，暂不添加索引
 	TenantID int `json:"tenant_id" gorm:"not null"`
 }
